[aotindutor] Forward fix a performance regression
Summary: Forward fix a performance regression caused by https://github.com/pytorch/pytorch/pull/110510. When a model is run once, all those kernel pointers are initialized and removing the if-nullptr check will cause those loadKernel be unnecessarily executed again when we rerun the foward function. Another way to do this is to codegen loadKernel in the initializer, which I may in a later PR.

ghstack-source-id: d2d5531df77c4e69c38e0e13c21278ca6943f0f0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/110800